// lib/type.ts
export interface MarketOrderRequest {
    acceptablePrice: string;
    accountId: string;
    allowAggregation: boolean;
    allowPartialMatching: boolean;
    chainId: number;
    expiration: string;
    id: string; // Signature
    marketId: string;
    nonce: string;
    referrerOrRelayer: string;
    settlementStrategyId: string;
    sizeDelta: string;
    trackingCode: string;
    reduceOnly: boolean;
}

/**
 * Payload structure used for signing a market order.
 */
export interface MarketOrderToSign {
    marketId: string;
    accountId: string;
    sizeDelta: string;
    settlementStrategyId: string;
    referrerOrRelayer: string;
    allowAggregation: boolean;
    allowPartialMatching: boolean;
    acceptablePrice: string;
    trackingCode: string;
    expiration: string;
    nonce: string;
    chainId: number;
    eoa: string;
    reduceOnly: boolean;
}